requirements-jax.txt
# JAX ecosystem for PyBird (JAX backend)
jax>=0.4.0
jaxlib>=0.4.0
interpax>=0.2.0
optax>=0.1.0
flax>=0.6.0
numpyro>=0.12.0
blackjax>=0.9.0
nautilus-sampler>=0.5.0 